refactores useAnswer, useUser, UseContext, UseComment
