+++ frontend/beCompliant/src/components/DeleteContextModal.tsx
@@ -10,7 +10,8 @@ import {
   Stack,
   Text,
 } from '@kvib/react';
-import { useDeleteContext } from '../hooks/useDeleteContext';
+
+import { useDeleteContext } from '../hooks/useFetchContext';
 
 type Props = {
   onOpen: () => void;
