+++ frontend/beCompliant/src/components/questionPage/RadioAnswer.tsx
@@ -1,7 +1,7 @@
 import { Text, RadioGroup, Radio, Stack, Flex } from '@kvib/react';
-import { useSubmitAnswers } from '../../hooks/useSubmitAnswers';
 import { Question, User } from '../../api/types';
 import { LastUpdated } from '../table/LastUpdated';
+import { useSubmitAnswers } from '../../hooks/useAnswers';
 
 type Props = {
   question: Question;
