+++ frontend/beCompliant/src/components/table/Comment.tsx
@@ -1,10 +1,10 @@
 import { Flex, IconButton, Text, Textarea, useDisclosure } from '@kvib/react';
 import { KeyboardEvent, useEffect, useRef, useState } from 'react';
-import { useSubmitComment } from '../../hooks/useSubmitComment';
 import { DeleteCommentModal } from './DeleteCommentModal';
 import { LastUpdated } from './LastUpdated';
 import { useCommentState } from './TableState';
 import { User } from '../../api/types';
+import { useSubmitComment } from '../../hooks/useComments';
 
 // Replace with type from api when the internal data model is implemented
 type Props = {
