--- frontend/beCompliant/src/hooks/useFetchAnswers.ts
@@ -1,25 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { apiConfig } from '../api/apiConfig';
-import { Answer } from '../api/types';
-import { axiosFetch } from '../api/Fetch';
-
-export function useFetchAnswers(contextId?: string) {
-  return useQuery({
-    queryKey: apiConfig.answers.queryKey(contextId!),
-    queryFn: () =>
-      axiosFetch<Answer[]>({
-        url: apiConfig.answers.url(contextId!),
-      }).then((response) => response.data),
-    select: formatAnswerData,
-    enabled: !!contextId,
-  });
-}
-
-function formatAnswerData(answers: Answer[]) {
-  return answers.map((answer: Answer) => {
-    return {
-      ...answer,
-      updated: new Date(answer.updated),
-    };
-  });
-}
