+++ frontend/beCompliant/src/pages/CreateContextPage.tsx
@@ -3,8 +3,8 @@ import { LockedCreateContextPage } from './LockedCreateContextPage';
 import { UnlockedCreateContextPage } from './UnlockedCreateContextPage';
 import { useFetchForms } from '../hooks/useFetchForms';
 import { Center, Heading, Icon } from '@kvib/react';
-import { useSubmitContext } from '../hooks/useSubmitContext';
 import { FormEvent, useCallback } from 'react';
+import { useSubmitContext } from '../hooks/useFetchContext';
 
 export const CreateContextPage = () => {
   const [search, setSearch] = useSearchParams();
