+++ frontend/beCompliant/src/pages/FrontPage.tsx
@@ -15,9 +15,11 @@ import {
 } from '@kvib/react';
 import { Link as ReactRouterLink } from 'react-router-dom';
 import { Page } from '../components/layout/Page';
-import { useFetchUserinfo } from '../hooks/useFetchUserinfo';
-import { useFetchTeamContexts } from '../hooks/useFetchTeamContexts';
-import { useFetchContext } from '../hooks/useFetchContext';
+import { useUser } from '../hooks/useUser';
+import {
+  useFetchContext,
+  useFetchTeamContexts,
+} from '../hooks/useFetchContext';
 import { useFetchForms } from '../hooks/useFetchForms';
 import { DeleteContextModal } from '../components/DeleteContextModal';
 import { apiConfig } from '../api/apiConfig';
@@ -28,7 +30,7 @@ const FrontPage = () => {
     data: userinfo,
     isPending: isUserinfoLoading,
     isError: isUserinfoError,
-  } = useFetchUserinfo();
+  } = useUser();
 
   const toast = useToast();
